Silence chpasswd during UID/GID change
Silence the `chpasswd` command on success to avoid the spurious
"chpasswd: password for 'git' changed" message. But still catch and
output the error message if the command fails.

During UID/GID setup the password change is a collateral of the setup
process, but not directly requested by the user. Password remains the
default one. Other `chpasswd` invocations are left as is since the
user did request that password to be changed.